financial-assistant/pyproject.toml
[project]
name = "financial-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.5",
    "aiohttp>=3.12.13",
    "anthropic>=0.55.0",
    "groq>=0.28.0",
    "openai>=1.91.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.3",
]
